requirements.txt
anthropic>=0.40.0
rdkit>=2024.3.1
python-dotenv>=1.0.0
pytest>=8.0.0
